Extract getGuildPreference to static method
