+++ src/commands/game_options/guessmode.ts
@@ -1,6 +1,5 @@
 import type BaseCommand from "../interfaces/base_command";
 import { IPCLogger } from "../../logger";
-import { getGuildPreference } from "../../helpers/game_utils";
 import {
     sendOptionsMessage,
     getDebugLogHeader,
@@ -14,6 +13,7 @@ import { GuessModeType } from "../../enums/option_types/guess_mode_type";
 import { DEFAULT_GUESS_MODE } from "../../constants";
 import LocalizationManager from "../../helpers/localization_manager";
 import Session from "../../structures/session";
+import GuildPreference from "../../structures/guild_preference";
 
 const logger = new IPCLogger("guessmode");
 
@@ -81,7 +81,9 @@ export default class GuessModeCommand implements BaseCommand {
     });
 
     call = async ({ message, parsedMessage }: CommandArgs): Promise<void> => {
-        const guildPreference = await getGuildPreference(message.guildID);
+        const guildPreference = await GuildPreference.getGuildPreference(
+            message.guildID
+        );
 
         if (parsedMessage.components.length === 0) {
             await guildPreference.reset(GameOption.GUESS_MODE_TYPE);
