+++ src/commands/game_options/ost.ts
@@ -2,7 +2,6 @@ import {
     getDebugLogHeader,
     sendOptionsMessage,
 } from "../../helpers/discord_utils";
-import { getGuildPreference } from "../../helpers/game_utils";
 import type BaseCommand from "../interfaces/base_command";
 import { IPCLogger } from "../../logger";
 import { GameOption } from "../../enums/game_option_name";
@@ -14,6 +13,7 @@ import { OstPreference } from "../../enums/option_types/ost_preference";
 import { DEFAULT_OST_PREFERENCE } from "../../constants";
 import LocalizationManager from "../../helpers/localization_manager";
 import Session from "../../structures/session";
+import GuildPreference from "../../structures/guild_preference";
 
 const logger = new IPCLogger("ost");
 
@@ -76,7 +76,9 @@ export default class OstCommand implements BaseCommand {
     });
 
     call = async ({ message, parsedMessage }: CommandArgs): Promise<void> => {
-        const guildPreference = await getGuildPreference(message.guildID);
+        const guildPreference = await GuildPreference.getGuildPreference(
+            message.guildID
+        );
 
         if (parsedMessage.components.length === 0) {
             await guildPreference.reset(GameOption.OST_PREFERENCE);
