+++ src/helpers/game_utils.ts
@@ -3,7 +3,7 @@ import { execSync } from "child_process";
 import dbContext from "../database_context";
 import State from "../state";
 import { IPCLogger } from "../logger";
-import GuildPreference from "../structures/guild_preference";
+import type GuildPreference from "../structures/guild_preference";
 import { cleanArtistName, cleanSongName } from "../structures/game_round";
 import SongSelector from "../structures/song_selector";
 import { LocaleType } from "../enums/locale_type";
@@ -85,42 +85,6 @@ export async function cleanupInactiveGameSessions(): Promise<void> {
             `Ended ${inactiveSessions} inactive game sessions out of ${totalSessions}`
         );
     }
-}
-
-/**
- * Gets or creates a GuildPreference
- * @param guildID - The Guild ID
- * @returns the correspond guild's GuildPreference
- */
-export async function getGuildPreference(
-    guildID: string
-): Promise<GuildPreference> {
-    const guildPreferences = await dbContext
-        .kmq("guilds")
-        .select("*")
-        .where("guild_id", "=", guildID);
-
-    if (guildPreferences.length === 0) {
-        const guildPreference = GuildPreference.fromGuild(guildID);
-        await dbContext
-            .kmq("guilds")
-            .insert({ guild_id: guildID, join_date: new Date() });
-        return guildPreference;
-    }
-
-    const gameOptionPairs = (
-        await dbContext
-            .kmq("game_options")
-            .select("*")
-            .where({ guild_id: guildID, client_id: process.env.BOT_CLIENT_ID })
-    )
-        .map((x) => ({ [x["option_name"]]: JSON.parse(x["option_value"]) }))
-        .reduce((total, curr) => Object.assign(total, curr), {});
-
-    return GuildPreference.fromGuild(
-        guildPreferences[0].guild_id,
-        gameOptionPairs
-    );
 }
 
 /**
